Refactor RPC Methods
